+++ src/Kiota.Builder/Writers/LanguageWriter.cs
@@ -10,6 +10,7 @@ using Kiota.Builder.Writers.Ruby;
 using Kiota.Builder.Writers.Shell;
 using Kiota.Builder.Writers.TypeScript;
 using Kiota.Builder.Writers.Php;
+using Kiota.Builder.Writers.Swift;
 
 namespace Kiota.Builder.Writers
 {
@@ -136,6 +137,7 @@ namespace Kiota.Builder.Writers
                 GenerationLanguage.PHP => new PhpWriter(outputPath, clientNamespaceName),
                 GenerationLanguage.Go => new GoWriter(outputPath, clientNamespaceName),
                 GenerationLanguage.Shell => new ShellWriter(outputPath, clientNamespaceName),
+                GenerationLanguage.Swift => new SwiftWriter(outputPath, clientNamespaceName),
                 _ => throw new InvalidEnumArgumentException($"{language} language currently not supported."),
             };
         }